+++ NEWS.md
@@ -1,5 +1,9 @@
 # AIscreenR 0.3.2.9000
 
+## Minor improvements
+* Adding the n_screened and n_missing variables to the key results in `screen_analyzer()`.
+
+
 # AIscreenR 0.3.2
 
 ## Minor improvements